Major changes in the project: the project is now a regular python package and extensive variation has been done on the MP pipeline in order to make it more general and flexibile

+++ de_novo_scilife/align.py
@@ -1,16 +1,14 @@
 import sys, os, yaml, glob
 import subprocess
-import common
 import string
 import sys
 import gzip
 import pandas as pd
-
+import common
 
 
 def _align_reads(global_config, sample_config, sorted_libraries_by_insert):
-    reference   = sample_config["reference"]
-    reference =  build_reference_bwa(global_config,reference)
+    reference =  build_reference_bwa(global_config, sample_config)
     #now prepare for alignment
     for library, libraryInfo in sorted_libraries_by_insert:
         read1       = libraryInfo["pair1"]
@@ -19,9 +17,10 @@ def _align_reads(global_config, sample_config, sorted_libraries_by_insert):
         insert      = libraryInfo["insert"]
         std         = libraryInfo["std"]
         threads     = 8
+        dryrun      = common.check_dryrun(sample_config)
         if "threads" in sample_config:
             threads  = sample_config["threads"]
-        libraryInfo["alignment"] = align_bwa_mem(global_config, read1, read2, reference, threads)
+        libraryInfo["alignment"] = align_bwa_mem(global_config, read1, read2, reference, threads, dryrun)
     return sorted_libraries_by_insert
 
 
@@ -35,6 +34,7 @@ def _merge_bam_files(global_config, sample_config, sorted_libraries_by_insert):
     elif not common.which("samtools"):
         sys.exit("error while trying to run  samtools: bwa not present in the path and not in global config, please make sure to install bwa properly")
 
+    numInserts = 0
     for library, libraryInfo in sorted_libraries_by_insert:
         read1       = libraryInfo["pair1"]
         read2       = libraryInfo["pair2"]
@@ -44,17 +44,23 @@ def _merge_bam_files(global_config, sample_config, sorted_libraries_by_insert):
         alignment   = libraryInfo["alignment"]
         if insert not in BAMfiles:
             BAMfiles[insert] = [alignment]
+            numInserts += 1
         else:
             BAMfiles[insert].append(alignment)
     
     BAMfilesMerged = {}
     for insert, insertGroup in BAMfiles.iteritems():
         dir_insert = "lib_{}".format(insert)
+        if numInserts == 1:
+            dir_insert = sample_config["output"]
         if not os.path.exists(dir_insert):
             os.makedirs(dir_insert)
         os.chdir(dir_insert)
         #check if file is already present
         bamMerged = "lib_{}.bam".format(insert)
+        if numInserts == 1:
+            bamMerged = "{}.bam".format(sample_config["output"])
+        
         if os.path.exists(bamMerged):
             BAMfilesMerged[insert] = os.path.abspath(bamMerged)
             os.chdir("..")
@@ -66,20 +72,22 @@ def _merge_bam_files(global_config, sample_config, sorted_libraries_by_insert):
             if  not returnValue == 0:
                 sys.exit("error, while soft linking {}".format(insertGroup[0]))
         else:
-            cl = [samtools, "merge"]
-            bamMerged = "lib_{}.bam".format(insert)
-            cl.append(bamMerged)
+            command = [samtools, "merge",bamMerged]
             for bamfile in insertGroup:
-                cl.append(bamfile)
-            returnValue = subprocess.call(cl)
-            if  not returnValue == 0:
-                sys.exit("error, while merging files {}".format(insertGroup))
+                command.append(bamfile)
+
+            common.print_command(command)
+            returnValue = 0
+            if not common.check_dryrun(sample_config):
+                returnValue = subprocess.call(command)
+                if  not returnValue == 0:
+                    sys.exit("error, while merging files {}".format(insertGroup))
         BAMfilesMerged[insert] = os.path.abspath(bamMerged)
         os.chdir("..")
     
     sorted_alignments_by_insert = []
     for key in sorted(BAMfilesMerged.iterkeys()):
-        sorted_alignments_by_insert.append([key, BAMfilesMerged[key]])
+        sorted_alignments_by_insert.append([key, BAMfilesMerged[key], dir_insert]) # memorise insert length, bam file, folder
     return sorted_alignments_by_insert
 
 
@@ -90,20 +98,20 @@ def picard_CGbias(global_config, sample_config, sorted_alignments_by_insert):
         picard = os.environ.get('PICARD_HOME')
     elif "picard" in global_config["Tools"]:
         picard = global_config["Tools"]["picard"]["bin"]
-    for library, BAMfile in sorted_alignments_by_insert:
-        working_dir = "lib_{}".format(library)
+    for library, BAMfile, working_dir in sorted_alignments_by_insert:
         os.chdir(working_dir)
         output_header = os.path.basename(BAMfile).split(".bam")[0]
         command= ["java", "-Xmx16g", "-XX:PermSize=2g", "-jar", os.path.join(picard, "CollectGcBiasMetrics.jar"),  "REFERENCE_SEQUENCE={}".format(sample_config["reference"]), "INPUT={}".format(BAMfile), \
         "OUTPUT={}.collectGcBias.txt".format(output_header), "CHART_OUTPUT={}.collectGcBias.pdf".format(output_header), "ASSUME_SORTED=true", "VALIDATION_STRINGENCY=LENIENT", "TMP_DIR=/scratch"]
-        print command
         returnValue = 0;
         if not os.path.exists("{}.collectGcBias.pdf".format(output_header)):
-            stdOut = open("collectGcBias.stdOut", "w")
-            stdErr = open("collectGcBias.stdErr", "w")
-            returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
-        if not returnValue == 0:
-            print "problem running collectGCBias"
+            common.print_command(command)
+            if not common.check_dryrun(sample_config):
+                stdOut = open("collectGcBias.stdOut", "w")
+                stdErr = open("collectGcBias.stdErr", "w")
+                returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
+                if not returnValue == 0:
+                    print "problem running collectGCBias"
         os.chdir("..")
     return sorted_alignments_by_insert
 
@@ -115,21 +123,21 @@ def picard_collectInsertSizeMetrics(global_config, sample_config, sorted_alignme
         picard = os.environ.get('PICARD_HOME')
     elif "picard" in global_config["Tools"]:
         picard = global_config["Tools"]["picard"]["bin"]
-    for library, BAMfile in sorted_alignments_by_insert:
-        working_dir = "lib_{}".format(library)
+    for library, BAMfile, working_dir in sorted_alignments_by_insert:
         os.chdir(working_dir)
         output_header = os.path.basename(BAMfile).split(".bam")[0]
         histWide = library * 2
         command= ["java", "-Xmx16g", "-XX:PermSize=2g", "-jar", os.path.join(picard, "CollectInsertSizeMetrics.jar"), "INPUT={}".format(BAMfile), "MINIMUM_PCT=0",
-        "HISTOGRAM_FILE={}.collectInsertSize.pdf".format(output_header), "OUTPUT={}.collectInsertSiz.txt".format(output_header), "HISTOGRAM_WIDTH={}".format(histWide), "VALIDATION_STRINGENCY=LENIENT", "TMP_DIR=/scratch"]
-        print command
+        "HISTOGRAM_FILE={}.collectInsertSize.pdf".format(output_header), "OUTPUT={}.collectInsertSize.txt".format(output_header), "HISTOGRAM_WIDTH={}".format(histWide), "VALIDATION_STRINGENCY=LENIENT", "TMP_DIR=/scratch"]
         returnValue = 0;
         if not os.path.exists("{}.collectInsertSize.pdf".format(output_header)):
-            stdOut = open("collectInsertSize.stdOut", "w")
-            stdErr = open("collectInsertSize.stdErr", "w")
-            returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
-        if not returnValue == 0:
-            print "problem running CollectInsertSizeMetrics"
+            common.print_command(command)
+            if not common.check_dryrun(sample_config):
+                stdOut = open("collectInsertSize.stdOut", "w")
+                stdErr = open("collectInsertSize.stdErr", "w")
+                returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
+                if not returnValue == 0:
+                    print "problem running CollectInsertSizeMetrics"
         os.chdir("..")
     return sorted_alignments_by_insert
 
@@ -140,26 +148,27 @@ def picard_markDuplicates(global_config, sample_config, sorted_alignments_by_ins
         picard = os.environ.get('PICARD_HOME')
     elif "picard" in global_config["Tools"]:
         picard = global_config["Tools"]["picard"]["bin"]
-    for library, BAMfile in sorted_alignments_by_insert:
-        working_dir = "lib_{}".format(library)
+    for library, BAMfile, working_dir in sorted_alignments_by_insert:
         os.chdir(working_dir)
         output_header = os.path.basename(BAMfile).split(".bam")[0]
         command= ["java", "-Xmx16g", "-XX:PermSize=3g", "-jar", os.path.join(picard, "MarkDuplicates.jar"), "INPUT={}".format(BAMfile), "OUTPUT={}_noDup.bam".format(output_header), \
           "METRICS_FILE={0}.markDuplicates.txt".format(output_header), "ASSUME_SORTED=true", "VALIDATION_STRINGENCY=LENIENT", "TMP_DIR=/scratch"]
-        print command
         returnValue = 0;
         if not os.path.exists("{}_noDup.bam".format(output_header)):
-            stdOut = open("removeDup.stdOut", "w")
-            stdErr = open("removeDup.stdErr", "w")
-            returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
-        if not returnValue == 0:
-            print "problem running MarkDuplicates"
+            common.print_command(command)
+            if not common.check_dryrun(sample_config):
+                stdOut = open("removeDup.stdOut", "w")
+                stdErr = open("removeDup.stdErr", "w")
+                returnValue = subprocess.call(command, stdout=stdOut, stderr=stdErr)
+                if not returnValue == 0:
+                    print "problem running MarkDuplicates"
         os.chdir("..")
     return sorted_alignments_by_insert
 
 
-def build_reference_bwa(global_config, reference):
+def build_reference_bwa(global_config, sample_config):
     #build the reference if not available
+    reference   = sample_config["reference"]
     program = "bwa"
     if "bwa" in global_config["Tools"]:
         program = global_config["Tools"]["bwa"]["bin"]
@@ -173,40 +182,36 @@ def build_reference_bwa(global_config, reference):
     current_dir           = os.getcwd()
     path_name, base_name  = os.path.split(reference)
     bwa_index_folder      = os.path.join(path_name, "bwa")
-
     #if needed create directory
     if not os.path.exists(bwa_index_folder):
         os.makedirs(bwa_index_folder)
     os.chdir(bwa_index_folder)
-
     # if needed soft link the reference
     if not os.path.exists(base_name):
         returnValue = subprocess.call(["ln", "-s", reference, base_name])
         if not returnValue == 0:
             sys.exit("error while trying to soft link reference sequence")
-
     reference = os.path.join(path_name, "bwa", base_name) # now I have a soflinked copy
     # now check if index alredy build or not
     if not os.path.exists("{}.bwt".format(reference)): # then create the index sequence
         bwa_stdOut = open("bwa_index.stdOut", "w")
         bwa_stdErr = open("bwa_index.stdErr", "w")
-        print "creating reference"
-        returnValue = subprocess.call([program, "index", reference], stdout=bwa_stdOut, stderr=bwa_stdErr)
-        if  not returnValue == 0:
-            sys.exit("error, while indexing reference file {} with bwa index".format(reference))
-
-    #extra control to avoid problem with nuexpected return value
+        command = [program, "index", reference]
+        common.print_command(command)
+        if not common.check_dryrun(sample_config):
+            returnValue = subprocess.call(command, stdout=bwa_stdOut, stderr=bwa_stdErr)
+            if  not returnValue == 0:
+                sys.exit("error, while indexing reference file {} with bwa index".format(reference))
+    #extra control to avoid problem with unexpected return value
     if not os.path.exists("{}.bwt".format(reference)):
         sys.exit("bwa index failed")
-
     os.chdir(current_dir)
-
     return reference
 
 
 
 
-def align_bwa_mem(global_config, read1, read2, reference, threads):
+def align_bwa_mem(global_config, read1, read2, reference, threads, dryrun):
     aligner = "bwa"
     if "bwa" in global_config["Tools"]:
         aligner = global_config["Tools"]["bwa"]["bin"]
@@ -219,10 +224,13 @@ def align_bwa_mem(global_config, read1, read2, reference, threads):
     elif not common.which("samtools"):
         sys.exit("error while trying to run  samtools: bwa not present in the path and not in global config, please make sure to install bwa properly")
 
-    # extract base name, assume the only dot is
-    libraryBase = os.path.basename(read1).split(".fastq")[0]
+    # extract base name
+    libraryBase = ""
     if read2:
-        libraryBase = "{}_{}".format(libraryBase.split("_")[0], libraryBase.split("_")[1])
+        libraryBase = os.path.basename(read1).split("_1.fastq")[0]
+    else:
+        libraryBase = os.path.basename(read1).split(".fastq")[0]
+
     if not os.path.exists(libraryBase):
         os.makedirs(libraryBase)
     os.chdir(libraryBase)
@@ -235,20 +243,7 @@ def align_bwa_mem(global_config, read1, read2, reference, threads):
         os.chdir("..")
         return BAMsorted
 
-    #if not os.path.exists(SAMMapped) and not os.path.exists(BAMunsorted):
-    #    with open(SAMMapped, "w") as fh:
-    #            stdErr = open("bwa.stdErr", "w")
-    #            returnValue = subprocess.call([aligner, "mem", "-M", "-t", "{}".format(threads), reference, read1, read2], stdout=fh, stderr=stdErr)
-    #            if  not returnValue == 0:
-    #                sys.exit("error, while aligning reads with bwa mem")
-    #
-    #if not os.path.exists(BAMunsorted):
-    #    with open(BAMunsorted, "w") as fh:
-    #            stdErr = open("bwa.stdErr", "w")
-    #            returnValue = subprocess.call([samtools, "view", "-S", "-F", "4", "-b", SAMMapped], stdout=fh, stderr=stdErr)
-    #           if  not returnValue == 0:
-    #                sys.exit("error, while while converting sam to bam")
-    
+
     bwa_mem_command       = [aligner, "mem", "-M", "-t", "{}".format(threads), reference, read1, read2]
     samtools_view_command = [samtools, "view", "-b", "-S",  "-u",  "-"]
 
@@ -256,14 +251,18 @@ def align_bwa_mem(global_config, read1, read2, reference, threads):
         command = "{} | {} > {}".format(" ".join(bwa_mem_command), " ".join(samtools_view_command), BAMunsorted)
         bwa_stdOut       = open("bwa.stdOut", "w")
         bwa_stdErr       = open("bwa.stdErr", "w")
-        subprocess.call(command, shell=True, stdout=bwa_stdOut, stderr=bwa_stdErr)
+        common.print_command(command)
+        if not dryrun:
+            subprocess.call(command, shell=True, stdout=bwa_stdOut, stderr=bwa_stdErr)
 
     samtools_sort_command = [samtools, "sort", "-@", "{}".format(threads), "-m" , "1G", BAMunsorted,  mappingBase]
     command = " ".join(samtools_sort_command)
     if not os.path.exists(BAMsorted):
         stdOut       = open("sam_sort.stdOut", "w")
         stdErr       = open("sam_sort.stdErr", "w")
-        subprocess.call(command, shell=True, stdout=stdOut, stderr=stdErr)
+        common.print_command(command)
+        if not dryrun:
+            subprocess.call(command, shell=True, stdout=stdOut, stderr=stdErr)
 
     if os.path.exists(BAMsorted) and os.path.exists(BAMunsorted)  :
         subprocess.call(["rm", BAMunsorted])
@@ -271,45 +270,6 @@ def align_bwa_mem(global_config, read1, read2, reference, threads):
     os.chdir("..")
     return BAMsorted
 
-
-def compl_rev(read_file):
-    originalHeader  = os.path.basename(read_file).split(".fastq")[0].split("_")
-    outputFileName  = "{}rc_{}_{}.fastq.gz".format(originalHeader[0],originalHeader[1],originalHeader[2])
-    outputFileName  = os.path.abspath(outputFileName)
-    if os.path.exists(outputFileName):
-        return outputFileName
-    
-    fd_in = ""
-    if ".gz" in read_file:
-        fd_in = gzip.open(filename=read_file, mode='rb')
-    else:
-        fd_in = open(filename, 'rb')
-
-
-    fd_out          = gzip.open(outputFileName, 'wb')
-    header = fd_in.readline().rstrip()
-    while header:
-        read    = fd_in.readline().rstrip()
-        comment = fd_in.readline().rstrip()
-        quality = fd_in.readline().rstrip()
-        read    = rc(read)
-        quality = quality[::-1]
-        fd_out.write("{}\n".format(header))
-        fd_out.write("{}\n".format(read))
-        fd_out.write("{}\n".format(comment))
-        fd_out.write("{}\n".format(quality))
-        header  = fd_in.readline().rstrip()
-    fd_in.close()
-    fd_out.close()
-    return outputFileName
-
-
-import string
-
-def rc(dna):
-    complements = string.maketrans('acgtnACGTN', 'tgcanTGCAN')
-    rcseq = dna.translate(complements)[::-1]
-    return rcseq
